+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -11,7 +11,7 @@ import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
 import org.apache.coyote.web.RequestMapping;
 import org.apache.coyote.web.request.HttpRequest;
-import org.apache.coyote.web.request.HttpRequestParser;
+import org.apache.coyote.support.HttpRequestParser;
 import org.apache.coyote.web.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
